+++ src/constants.ts
@@ -3,6 +3,7 @@ export const regexes = {
   hex: "^(?:0x|#)?[0-9A-Fa-f]+$",
   /** https://regex101.com/r/KTruwH/1 */
   decimal: "^[0-9]+$",
+  binary: "^[0-1]+$",
   whitespace: "\\s",
   newline: "\\n",
   character: "[a-zA-Z0-9]",