+++ src/extension.ts
@@ -2,12 +2,13 @@ import * as vscode from "vscode";
 import { ConvertCallback, MatcherCallback } from "./types";
 import { forEachLineIn, forEachWordIn } from "./lib";
 import {
+  binaryStringToHexString,
   decimalStringToHexString,
   hexStringToASCIIString,
   hexStringToBinaryString,
   hexStringToDecimalString,
 } from "./converters";
-import { isDecimalString, isHexString } from "./matchers";
+import { isBinaryString, isDecimalString, isHexString } from "./matchers";
 import { insertLineComment } from "./comments";
 
 export function comment(
@@ -140,6 +141,10 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand("hex-multi-converter.decimalToHex", () => {
       const editor = vscode.window.activeTextEditor;
       replace(editor, decimalStringToHexString, isDecimalString);
+    }),
+    vscode.commands.registerCommand("hex-multi-converter.binaryToHex", () => {
+      const editor = vscode.window.activeTextEditor;
+      replace(editor, binaryStringToHexString, isBinaryString);
     }),
   );
 }