+++ src/matchers.ts
@@ -11,3 +11,9 @@ const decimalRegExp = new RegExp(regexes.decimal);
 export function isDecimalString(value: string): boolean {
   return decimalRegExp.test(value);
 }
+
+const binaryRegExp = new RegExp(regexes.binary);
+
+export function isBinaryString(value: string): boolean {
+  return binaryRegExp.test(value);
+}